Добавил ToString в Book

+++ Shoping/IProduct.cs
@@ -7,7 +7,7 @@ namespace Shopping
     interface IProduct
     {
         string Name { get; set; }
-        float Price { get; set; }
+        double Price { get; set; }
         int Rating { get; set; }
         int Id { get; set; }
     }
